Finished kitchen sql part
i worked on the procedures for the kitchen. It was hard becuase once the burger was created we didnt put it in any table so i had to fix bunch of tables including the table orders so insure that once the order is created it actually went somewhere but i didnt. The next person that works on the project just has to connect the sql code to the javascript part.