New setup
We have discussed to make a new simple python project instead of the old one because we couldn't figure out how to use docker on the old project so i created a basic setup for our python project. I made a new directory called my_burger theres 2 folders inside of it one for kitchen view containing all the kitchen view files and dockerfile and i made a new order folder that contains all the order files and the order dockerfile and at the root of out directory i made  docker compose file.